pyproject.toml
[project]
name = "finicial-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "finnhub-python>=2.4.22",
    "icecream>=2.1.4",
    "pandas>=2.2.3",
    "pendulum>=3.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "ratelimit>=2.2.1",
]
